Fix error when rendering chart with helm version 2
ArgoCD selects the helm version to use based on the api version of the
helm chart. Because this chart uses api v1 this means argocd will try
and render the helm chart using helm v2.

In helm v2 there seems to be a bug where the trimming of trailing space
does too much trimming and results in a line that reads as:

'---apiVersion': rbac.authorization.k8s.io/v1 which is invalid.

Therefore argocd refuses to sync the helm chart. Helm v3 does not
have this bug present. However removing the trailing trim operator does
not seem to have a negative effect on the rendering of the chart and
fixes helm v2.